refactor worker code
